Refactored profiler storage tests to share some code.
